Inactive content in Foro will not show up, plus Admin and Foro fixes
Added post_confirm_delete.html for later implementation.
Fixed a bug where searching for a Post or a Comment would crash the Admin page due to search_fields problems.
Fixed a problem of redirection to Foro.html after Deletion or Update of Comments.
Changed the filtering of content in Foro to views.py instead of the html templates.
Changed Comments so that now they can be created on the same page of the Post.